code/multiagent/pyproject.toml
[project]
name = "multiagent"
version = "0.1.0"
description = "agents over the A2A protocol."
readme = "README.md"
requires-python = ">=3.10,<3.13"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
a2a-sdk = ">=0.2.5"
httpx = ">=0.28.1"
uvicorn = ">=0.34.2"
google-adk = ">=1.0.0"
google-genai = ">=1.9.0"
python-dotenv = ">=1.1.0"
pydantic = ">=2.10.6"
litellm = ">=1.72.4"
mcp = ">=1.9.2"
openai = ">=1.84.0"
sseclient-py = ">=1.8.0"
graphrag = ">=2.4.0"
langchain = "==0.2.16"
langchain-core = "==0.2.38"
langchain-community = "==0.2.16"
venus-sdk = "==1.3.26"
trag = ">=0.0.52"
trag-client = "0.6.0"
trag-langchain = "0.0.31"

[[tool.poetry.source]]
name = "tencent"
url = "https://mirrors.tencent.com/pypi/simple/"
priority = "primary"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"